fix: remove all compilation warnings (177 → 11)
- Remove ~100 unused `use oxc_span::GetSpan` imports
- Remove ~15 unused AST type imports (Expression, AstType, AstKind, etc.)
- Remove unnecessary parentheses and unreachable match arms
- Remove isolated dead functions (path_looks_like_auth, strip_quotes,
  args_source_has_body, collect_chain_methods, has_force_true)
- Apply clippy auto-fixes (while let loops, if-let collapse, find_map
  simplification, unused lifetime)

Remaining 11 warnings are systemic dead code from tree-sitter backends
migrated to oxc but not yet removed.

+++ src/rules/jsdoc_missing_example/oxc_typescript.rs
@@ -37,7 +37,7 @@ impl OxcCheck for Check {
             return;
         };
 
-        let export_start = export.span.start as u32;
+        let export_start = export.span.start;
 
         // Find a JSDoc comment preceding this export.
         let Some(jsdoc_text) = find_jsdoc_above(semantic, ctx.source, export_start) else {
